requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
Flask==0.12.3
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.19
pymessenger==0.0.7.0
python-telegram-bot==11.1.0
requests==2.21.0
requests-toolbelt==0.8.0
selenium==3.141.0
six==1.12.0
soupsieve==1.7.1
urllib3==1.24.1
Werkzeug==0.14.1
